Move switchType to hidden header
